Move GPU utils to utils_notebook.py
