+++ batchflow/__init__.py
@@ -18,9 +18,10 @@ from .dsindex import DatasetIndex, FilesIndex
 from .decorators import action, inbatch_parallel, parallel, any_action_failed, mjit, deprecated, apply_parallel
 from .exceptions import SkipBatchException, EmptyBatchSequence, StopPipeline
 from .sampler import Sampler, ConstantSampler, NumpySampler, HistoSampler, ScipySampler
-from .utils import save_data_to, read_data_from, get_available_gpus, set_gpus
+from .utils import save_data_to, read_data_from
 from .utils_random import make_rng, make_seed_sequence, spawn_seed_sequence
-from .utils_notebook import in_notebook, get_notebook_path, get_notebook_name, pylint_notebook
+from .utils_notebook import in_notebook, get_notebook_path, get_notebook_name, pylint_notebook,\
+                            get_available_gpus, set_gpus
 
 
 __version__ = '0.5.0beta3'
