+++ batchflow/utils_notebook.py
@@ -3,7 +3,9 @@ import os
 import re
 import json
 
-# Additionally imports 'requests`, 'ipykernel`, `notebook`, `nbconvert` and `pylint`, if needed
+import numpy as np
+
+# Additionally imports 'requests`, 'ipykernel`, `notebook`, `nbconvert`, `pylint` and `nvidia_smi`, if needed
 
 
 def in_notebook():
@@ -183,3 +185,77 @@ def pylint_notebook(path=None, options='', printer=print, ignore_comments=True, 
     if return_report:
         return '\n'.join(report_)
     return 0
+
+
+def get_available_gpus(n=1, min_free_memory=0.9, max_processes=2, verbose=False):
+    """ Select `n` gpus from available and free devices.
+
+    Parameters
+    ----------
+    n : int
+        Number of devices to select.
+    min_free_memory : float
+        Minimum percentage of free memory on a device to consider it free.
+    max_processes : int
+        Maximum amount of computed processes on a device to consider it free.
+    verbose : bool
+        Whether to show individual device information.
+
+    Returns
+    -------
+    List with indices of availble GPUs
+    """
+    try:
+        import nvidia_smi
+    except ImportError as exception:
+        raise ImportError('Install Python interface for nvidia_smi') from exception
+
+    nvidia_smi.nvmlInit()
+    n_devices = nvidia_smi.nvmlDeviceGetCount()
+
+    available_devices, memory_usage = [], []
+    for i in range(n_devices):
+        handle = nvidia_smi.nvmlDeviceGetHandleByIndex(i)
+        info = nvidia_smi.nvmlDeviceGetMemoryInfo(handle)
+
+        fraction_free = info.free / info.total
+        num_processes = len(nvidia_smi.nvmlDeviceGetComputeRunningProcesses(handle))
+
+        consider_available = (fraction_free > min_free_memory) & (num_processes <= max_processes)
+        if consider_available:
+            available_devices.append(i)
+            memory_usage.append(fraction_free)
+
+        if verbose:
+            print(f'Device {i} | Free memory: {fraction_free:4.2f} | '
+                  f'Number of running processes: {num_processes:>2} | Free: {consider_available}')
+
+    if len(available_devices) < n:
+        raise ValueError(f'Not enough free devices: requested {n}, found {len(available_devices)}')
+    available_devices = np.array(available_devices)[np.argsort(memory_usage)[::-1]]
+    return sorted(available_devices[:n])
+
+def set_gpus(n=1, min_free_memory=0.9, max_processes=2, verbose=False):
+    """ Set the `CUDA_VISIBLE_DEVICES` variable to `n` available devices.
+
+    Parameters
+    ----------
+    n : int
+        Number of devices to select.
+    min_free_memory : float
+        Minimum percentage of free memory on a device to consider it free.
+    max_processes : int
+        Maximum amount of computed processes on a device to consider it free.
+    verbose : bool
+        Whether to show individual device information.
+    """
+    if 'CUDA_VISIBLE_DEVICES' in os.environ.keys():
+        raise ValueError('`CUDA_VISIBLE_DEVICES` is already set!')
+
+    devices = get_available_gpus(n=n, min_free_memory=min_free_memory, max_processes=max_processes, verbose=verbose)
+    str_devices = ','.join(str(i) for i in devices)
+    os.environ['CUDA_VISIBLE_DEVICES'] = str_devices
+
+    newline = "\n" if verbose else ""
+    print(f'{newline}`CUDA_VISIBLE_DEVICES` set to "{str_devices}"')
+    return devices
